icu-system/rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
langchain==0.1.0
langchain-community==0.0.10
groq>=0.9.0
httpx>=0.27.0
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0
